re ordered everything, but need to fix links
