+++ crates/store/re_log_encoding/src/protobuf_conversions.rs
@@ -1,4 +1,6 @@
-impl From<re_protos::log_msg::v0::Compression> for crate::Compression {
+use crate::codec::rrd::Compression;
+
+impl From<re_protos::log_msg::v0::Compression> for Compression {
     fn from(value: re_protos::log_msg::v0::Compression) -> Self {
         match value {
             re_protos::log_msg::v0::Compression::None => Self::Off,
@@ -7,11 +9,11 @@ impl From<re_protos::log_msg::v0::Compression> for crate::Compression {
     }
 }
 
-impl From<crate::Compression> for re_protos::log_msg::v0::Compression {
-    fn from(value: crate::Compression) -> Self {
+impl From<Compression> for re_protos::log_msg::v0::Compression {
+    fn from(value: Compression) -> Self {
         match value {
-            crate::Compression::Off => Self::None,
-            crate::Compression::LZ4 => Self::Lz4,
+            Compression::Off => Self::None,
+            Compression::LZ4 => Self::Lz4,
         }
     }
 }
@@ -19,9 +21,8 @@ impl From<crate::Compression> for re_protos::log_msg::v0::Compression {
 #[cfg(feature = "decoder")]
 pub fn log_msg_from_proto(
     message: re_protos::log_msg::v0::LogMsg,
-) -> Result<re_log_types::LogMsg, crate::decoder::DecodeError> {
-    use crate::codec::{arrow::decode_arrow, CodecError};
-    use crate::decoder::DecodeError;
+) -> Result<re_log_types::LogMsg, crate::codec::DecodeError> {
+    use crate::codec::{arrow::decode_arrow, CodecError, DecodeError};
     use re_protos::{
         log_msg::v0::{log_msg::Msg, Encoding},
         missing_field,
@@ -66,8 +67,8 @@ pub fn log_msg_from_proto(
 #[cfg(feature = "encoder")]
 pub fn log_msg_to_proto(
     message: re_log_types::LogMsg,
-    compression: crate::Compression,
-) -> Result<re_protos::log_msg::v0::LogMsg, crate::encoder::EncodeError> {
+    compression: Compression,
+) -> Result<re_protos::log_msg::v0::LogMsg, crate::codec::EncodeError> {
     use crate::codec::arrow::encode_arrow;
     use re_protos::log_msg::v0::{
         ArrowMsg, BlueprintActivationCommand, LogMsg as ProtoLogMsg, SetStoreInfo,
@@ -87,8 +88,8 @@ pub fn log_msg_to_proto(
             let arrow_msg = ArrowMsg {
                 store_id: Some(store_id.into()),
                 compression: match compression {
-                    crate::Compression::Off => re_protos::log_msg::v0::Compression::None as i32,
-                    crate::Compression::LZ4 => re_protos::log_msg::v0::Compression::Lz4 as i32,
+                    Compression::Off => re_protos::log_msg::v0::Compression::None as i32,
+                    Compression::LZ4 => re_protos::log_msg::v0::Compression::Lz4 as i32,
                 },
                 uncompressed_size: payload.uncompressed_size as i32,
                 encoding: re_protos::log_msg::v0::Encoding::ArrowIpc as i32,
